requirements.txt
anyio==3.7.1
arabic-reshaper==3.0.0
asn1crypto==1.5.1
async-generator==1.10
attrs==23.1.0
autopep8==2.0.4
bcrypt==4.0.1
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-spinner==0.1.10
colorama==0.4.6
croniter==1.4.1
cryptography==41.0.7
cssselect2==0.7.0
exceptiongroup==1.1.3
fastapi==0.103.1
fastapi-login==1.9.1
fastapi-utilities==0.2.0
freetype-py==2.3.0
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.3
httpx==0.26.0
idna==3.4
iniconfig==2.0.0
lxml==5.1.0
oscrypto==1.3.0
outcome==1.2.0
packaging==23.2
passlib==1.7.4
pillow==10.2.0
pluggy==1.4.0
psycopg2==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pydantic==1.10.12
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pypng==0.20220715.0
pytest==8.0.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-multipart==0.0.6
PyYAML==6.0.1
qrcode==7.4.2
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
starlette==0.27.0
tinycss2==1.2.1
tomli==2.0.1
trio==0.21.0
typing_extensions==4.7.1
tzdata==2023.4
tzlocal==5.2
uritools==4.0.2
urllib3==2.1.0
uvicorn==0.23.2
webencodings==0.5.1
